final commit

+++ database/migrations/2021_04_03_051453_create_product_sales_table.php
@@ -19,7 +19,7 @@ class CreateProductSalesTable extends Migration
             $table->bigInteger('product_id');
             $table->decimal('price',10,2);
             $table->integer('quantity');
-            $table->integer('total');
+            $table->decimal('total',10,2);
             $table->timestamps();
         });
     }
